Added team command

+++ src/main/kotlin/dev/triumphteam/contest/listeners/VotingListener.kt
@@ -28,6 +28,7 @@ fun JdaApplication.voting() {
 
     on<GuildMessageReceivedEvent> {
         if (!message.contentDisplay.startsWith("!vote")) return@on
+        if (config[Settings.STARTED]) return@on
         message.delete().queue()
 
         channel.sendMessage(
